Refactor to use a centrally defined marker hashtag
