+++ sdk/eventhub/event-hubs/test/eventdata.spec.ts
@@ -3,6 +3,7 @@
 
 import chai from "chai";
 chai.should();
+import { runnableInBrowserMarker } from "./utils/testUtils";
 
 import { EventData, Message } from "../src";
 
@@ -32,7 +33,7 @@ const testMessage: Message = {
 const testEventData = EventData.fromAmqpMessage(testMessage);
 const messageFromED = EventData.toAmqpMessage(testEventData);
 
-describe("EventData #RunInBrowser", function(): void {
+describe(`EventData ${runnableInBrowserMarker}`, function(): void {
   describe("fromAmqpMessage", function(): void {
     it("populates annotations with the message annotations", function(): void {
       testEventData.annotations!.should.equal(testAnnotations);
