+++ sdk/eventhub/event-hubs/test/iothub.spec.ts
@@ -8,10 +8,10 @@ chai.use(chaiAsPromised);
 import debugModule from "debug";
 const debug = debugModule("azure:event-hubs:iothub-spec");
 import { EventHubClient } from "../src";
-import { EnvVarKeys, getEnvVars } from "./utils/envVarUtils";
+import { EnvVarKeys, getEnvVars } from "./utils/testUtils";
 const env = getEnvVars();
 
-describe("EventHub Client with iothub connection string #NotYetRunInBrowser", function(): void {
+describe("EventHub Client with iothub connection string", function(): void {
   const service = { connectionString: env[EnvVarKeys.IOTHUB_CONNECTION_STRING] };
   let client: EventHubClient;
   before("validate environment", async function(): Promise<void> {
