+++ sdk/eventhub/event-hubs/test/sender.spec.ts
@@ -8,10 +8,10 @@ chai.use(chaiAsPromised);
 import debugModule from "debug";
 const debug = debugModule("azure:event-hubs:sender-spec");
 import { EventHubClient, EventData } from "../src";
-import { EnvVarKeys, getEnvVars } from "./utils/envVarUtils";
+import { EnvVarKeys, getEnvVars, runnableInBrowserMarker } from "./utils/testUtils";
 const env = getEnvVars();
 
-describe("EventHub Sender #RunInBrowser", function(): void {
+describe(`EventHub Sender ${runnableInBrowserMarker}`, function(): void {
   const service = {
     connectionString: env[EnvVarKeys.EVENTHUB_CONNECTION_STRING],
     path: env[EnvVarKeys.EVENTHUB_NAME]
